13_model_settings/pyproject.toml
[project]
name = "13-model-settings"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.4",
]
